infrastructure/testdriver/execute/test_async.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>TestDriver execute method (async)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
setup({single_test: true})
onload = async () => {
    let uuid = "{{uuid()}}";
    let url = "https://{{host}}:{{ports[https][0]}}/infrastructure/testdriver/execute/child_slow.html?uuid=" + uuid;
    win = window.open(url);
    let content;
    while (true) {
        content = await test_driver.execute("let callback = arguments[arguments.length - 1]; if (location.href == 'about:blank') {return null}; setTimeout(() => callback(document.body.textContent), 1000)", true, uuid);
        if (content !== null) {
            break
        }
    }
    assert_equals(content.trim(), "PASS");
    done();
}
</script>
